pyproject.toml
[project]
name = "fdua-competition"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.15",
    "langchain-chroma>=0.2.1",
    "langchain-community>=0.3.15",
    "langchain-openai>=0.3.2",
    "langsmith>=0.3.1",
    "openai>=1.60.1",
    "pypdfium2>=4.30.1",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "ipython>=8.31.0",
    "isort>=5.13.2",
    "pandas>=2.2.3",
    "pytest>=8.3.4",
    "ruff>=0.9.3",
    "tabulate>=0.9.0",
]

[tool.ruff]
line-length = 127
ignore = ["E402"]